requirements.txt
--extra-index-url="https://packages.dea.ga.gov.au"

boto3
click
odc_aws
urlpath
# For tests. Todo, only install these in the image when we're testin
pytest
moto
